added separate schedules in tabs
