+++ package.json
@@ -19,8 +19,7 @@
     "@vitejs/plugin-react": "^4.3.1",
     "date-fns": "^4.1.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "tailwind": "^4.0.0"
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
